change int to big int

+++ app/words_game/models.py
@@ -1,14 +1,15 @@
 
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship, MappedAsDataclass
+from typing_extensions import Annotated
 
-from app.store.database.sqlalchemy_base import DB
+from app.store.database.sqlalchemy_base import DB, bigint
 
 
 class User(MappedAsDataclass, DB):
     __tablename__ = 'users'
 
-    id: Mapped[int] = mapped_column(primary_key=True)
+    id: Mapped[bigint] = mapped_column(primary_key=True)
     username: Mapped[str] = mapped_column(nullable=False)
 
 
@@ -17,13 +18,13 @@ class GameSession(MappedAsDataclass, DB):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     game_type: Mapped[str] = mapped_column(nullable=False)
-    chat_id: Mapped[int] = mapped_column(nullable=False)
+    chat_id: Mapped[bigint] = mapped_column(nullable=False)
     words: Mapped[str] = mapped_column(nullable=True)
-    next_user_id: Mapped[int] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'), nullable=True)
+    next_user_id: Mapped[bigint] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'), nullable=True)
     next_user: Mapped[User] = relationship(User, lazy='joined', foreign_keys=[next_user_id])
-    user_id: Mapped[int] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'), nullable=True)
+    user_id: Mapped[bigint] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'), nullable=True)
     user: Mapped[User] = relationship(User, lazy='joined', foreign_keys=[user_id])
-    winner_id: Mapped[int] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'), nullable=True)
+    winner_id: Mapped[bigint] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'), nullable=True)
     winner: Mapped[User] = relationship(User, lazy='joined', foreign_keys=[winner_id])
     status: Mapped[bool] = mapped_column(nullable=False, default=False)
     next_start_letter: Mapped[str] = mapped_column(default=None, nullable=True)
@@ -35,7 +36,7 @@ class Team(MappedAsDataclass, DB):
     id: Mapped[int] = mapped_column(primary_key=True)
 
     game_sessions_id: Mapped[int] = mapped_column(ForeignKey('game_sessions.id', ondelete='CASCADE'))
-    player_id: Mapped[int] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'))
+    player_id: Mapped[bigint] = mapped_column(ForeignKey('users.id', ondelete='CASCADE'))
     game_session: Mapped[GameSession] = relationship(GameSession, backref='teams')
     players: Mapped[list[User]] = relationship(User, backref='teams')
     life: Mapped[int] = mapped_column(default=3)
